% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{contrast}
\alias{contrast}
\title{Contrast}
\usage{
contrast(input_img, intensity = 5, output_img)
}
\arguments{
\item{input_img}{string, path of the input .png image}

\item{intensity}{int, intensity of contrast enhancement, between 0 and 10, defaults to 5}

\item{output_img}{string, path of the output .png image}
}
\value{
a png image at the specificed out path
}
\description{
Contrast
}
\examples{
contrast("./input.png", 5, "./output.png")
}
