% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpen.R
\name{sharpen}
\alias{sharpen}
\title{Sharpen}
\usage{
sharpen(input_img, intensity = 5, display = F, output_img = "")
}
\arguments{
\item{input_img}{string, path of the input .png image}

\item{intensity}{int, intensity of sharpness enhancement, between 0 and 10, defaults to 5}

\item{display}{bool, if TRUE, output will be displayed. Defaults to FALSE.}

\item{output_img}{string, path of the output .png image}
}
\value{
a png image at the specificed out path
}
\description{
Sharpens an image with given intensity specification.
This function will use the unsharp mask approach to sharpen images.
Works by enhancing the edges of an image.
}
\examples{
\dontrun{
sharpen("./input.png", 5, FALSE, "./output.png")
}
}
