% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{contrast}
\alias{contrast}
\title{Adjust contrast of image}
\usage{
contrast(input_img, intensity = 5, display = F, output_img = "")
}
\arguments{
\item{input_img}{string, path of the input image}

\item{intensity}{int, intensity of contrast enhancement, between 0 and 10, defaults to 5}

\item{display}{bool, if TRUE, output will be displayed. Defaults to FALSE.}

\item{output_img}{string, path of the output image}
}
\value{
an image at the specificed output path
}
\description{
Adjust the contrast of an image, given the intensity specification.
Works by making dark pixels much darker, and light pixels only slighly darker.
}
\examples{
\dontrun{
contrast("./input.png", 5, FALSE, "./output.png")
}
}
