% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{contrast}
\alias{contrast}
\title{Contrast}
\usage{
contrast(input_img, intensity, output_img)
}
\arguments{
input_img - path to load .png image {string}
intensity - desired contrast intensity [0,10] {integer}
output_img - path to save .png image with adjusted contrast {string}
}
\item{output_img}{
string path of output image
}
\description{
Created on February 9, 2019
Author: George J. J. Wu
Contrast, make dark pixels much darker, and bright pixels slightly darker, returns a .png image.
}
\example{
contrast("../img/image.png", 2, "../img/contrast_image.png")
}