% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpen.R
\name{sharpen}
\alias{sharpen}
\title{Sharpen}
\usage{
sharpen(input_img, intensity = 5, output_img)
}
\arguments{
\item{input_img}{string, path of the input .png image}

\item{intensity}{int, intensity of sharpness enhancement, between 0 and 10, defaults to 5}

\item{output_img}{string, path of the output .png image}
}
\value{
a png image at the specificed out path
}
\description{
Sharpen
}
\examples{
sharpen("./input.png", 5, "./output.png")
}
