% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibrance.R
\name{vibrance}
\alias{vibrance}
\title{Vibrance}
\usage{
vibrance(input_img, intensity = 5, display = F, output_img = "")
}
\arguments{
\item{input_img}{string, path of the input .png image}

\item{intensity}{int, intensity of vibrance enhancement, between -10 and 10, defaults to 5}

\item{display}{bool, if TRUE, output will be displayed. Defaults to FALSE.}

\item{output_img}{string, path of the output .png image}
}
\value{
a png image at the specificed out path
}
\description{
Vibrance
}
\examples{
\dontrun{
vibrance("./input.png", 5, FALSE, "./output.png")
}
}
